requirements.txt

Speech-Recognition
azure-cognitiveservices-speech
os
random
pyaudio
pydub
audio_segment
pyttsx3
flask
python-multipart
fastapi
ping
json
requests
regex

